+++ app/src/main/java/io/vn/nguyenduck/blocktopograph/activity/MainActivity.java
@@ -1,10 +1,5 @@
 package io.vn.nguyenduck.blocktopograph.activity;
 
-import static io.vn.nguyenduck.blocktopograph.Constants.MINECRAFT_APP_ID;
-import static io.vn.nguyenduck.blocktopograph.utils.Utils.buildAndroidDataDir;
-import static io.vn.nguyenduck.blocktopograph.utils.Utils.isAndroid11Up;
-
-import android.annotation.SuppressLint;
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
@@ -17,9 +12,7 @@ import androidx.navigation.ui.NavigationUI;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 import io.vn.nguyenduck.blocktopograph.R;
-import io.vn.nguyenduck.blocktopograph.file.BFile;
 
-@SuppressLint("SdCardPath")
 public class MainActivity extends AppCompatActivity {
 
     @Override
@@ -34,23 +27,5 @@ public class MainActivity extends AppCompatActivity {
         NavController navController = NavHostFragment.findNavController(contentView);
 
         NavigationUI.setupWithNavController(bottomNavigationView, navController);
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        if (isAndroid11Up()) {
-            BFile f = new BFile(buildAndroidDataDir(MINECRAFT_APP_ID) + "/games");
-            f.copyTo(new BFile("/sdcard"));
-        }
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        if (isAndroid11Up()) {
-            BFile f = new BFile("/sdcard/games");
-            f.copyTo(new BFile(buildAndroidDataDir(MINECRAFT_APP_ID)));
-        }
     }
 }