+++ app/src/main/java/io/vn/nguyenduck/blocktopograph/setting/ASetting.java
@@ -1,0 +1,20 @@
+package io.vn.nguyenduck.blocktopograph.setting;
+
+public abstract class ASetting {
+
+    public Object value = getDefaultValue();
+
+    public abstract String getKey();
+
+    public abstract String getCategory();
+
+    public abstract String getName();
+
+    public abstract String getDescription();
+
+    public abstract Object getDefaultValue();
+
+    public Class<?> getType() {
+        return getDefaultValue().getClass();
+    }
+}